requirements.txt
# genix_alz/requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.1
fpdf2==2.7.5
joblib==1.3.2
